requirements.txt
strands-agents
strands-agents-tools
anthropic
python-dotenv
fastapi
uvicorn[standard]
ag-ui-protocol
plotly
pandas
